@Hongbosherlock 2025-07-13
> @Hongbosherlock Maybe check on this as well, before you prepare the PR for nixl.

Got it. My bad, I didn't consider the DP attention scenario for non-deepseek models earlier.